nibble v0.0.1 NibbleWeb.Router.Helpers

Module with named helpers generated from NibbleWeb.Router.

Functions

book_path(conn_or_endpoint, action)

book_path(conn_or_endpoint, action, params)

book_path(conn_or_endpoint, action, id, params)

book_url(conn_or_endpoint, action)

book_url(conn_or_endpoint, action, params)

book_url(conn_or_endpoint, action, id, params)

page_path(conn_or_endpoint, action)

page_path(conn_or_endpoint, action, params)

page_path(conn_or_endpoint, action, path, params)

page_url(conn_or_endpoint, action)

page_url(conn_or_endpoint, action, params)

page_url(conn_or_endpoint, action, path, params)

path(data, path)

Generates the path information including any necessary prefix.

place_path(conn_or_endpoint, action)

place_path(conn_or_endpoint, action, params)

place_path(conn_or_endpoint, action, id, params)

place_url(conn_or_endpoint, action)

place_url(conn_or_endpoint, action, params)

place_url(conn_or_endpoint, action, id, params)

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.

user_path(conn_or_endpoint, action)

user_path(conn_or_endpoint, action, params)

user_path(conn_or_endpoint, action, id, params)

user_url(conn_or_endpoint, action)

user_url(conn_or_endpoint, action, params)

user_url(conn_or_endpoint, action, id, params)